+++ CHANGES.rst
@@ -15,6 +15,7 @@ Changelog
 - Make the searchlastdecisions meetings search able to display decisions in the future
 - Do not display the 'review_state' columns in contacts dashboard displaying organizations, it is always 'active', we use the 'selected in plonegroup' column information instead
 - Fixed migration of MeetingUsers, do not fail if a MeetingUser was deleted and initialize MeetingConfig.orderedContacts and MeetingConfig.orderedItemInitiators correctly
+- Added possibility to use a DashboardPODTemplate added into the contacts directory on contacts dashboards
 
 4.1 (2019-08-23)
 ----------------